+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- CS-554 - Extended TrialSystem model with effective knowledge attribute and create backend import page
 - CS-555 - Added google calendar tab to Association, District and Team
 - CS-475 - Added option to backend scout list to (soft)delete scout with personal data
 - CS-520 - Modification and optimization of grid display and the design of the filtering